completar ejercicio 2_2

+++ src/main/java/edu/badpals/empresa/DatabaseManager.java
@@ -73,19 +73,82 @@ public class DatabaseManager {
 
         }catch (SQLException e){
             System.out.println("Error al crear el departamento.");
-            e.printStackTrace();
         }
     }
 
     public static void borrarEmpleadoProyecto(String nss, int numProyecto){
         try{
+            ResultSet chkEmpleado = statement.executeQuery("select 1 from empregado where NSS = '"+nss+"'");
+            if(!chkEmpleado.next()){
+                System.out.println("El empleado introducido no existe");
+                return;
+            }
+            chkEmpleado.close();
+
+            ResultSet chkProyecto = statement.executeQuery("select 1 from proxecto where NUM_PROXECTO = "+numProyecto+"");
+            if(!chkProyecto.next()){
+                System.out.println("El número de proyecto introducido no existe");
+                return;
+            }
+            chkProyecto.close();
+
             int numFilasAfectadas = statement.executeUpdate("delete from empregado_proxecto where NSS_EMPREGADO = '"+nss+"' and NUM_PROXECTO = "+numProyecto+";");
             System.out.println("Registro de proyecto borrado. Filas afectadas: "+numFilasAfectadas);
         }catch (SQLException e){
             System.out.println("Error al borrar el registro de proyecto");
+        }
+    }
+
+    public static void getEmpleadosLocalidad(String localidad){
+        try{
+            ResultSet rs = statement.executeQuery("select\n" +
+                    "NOME,APELIDO_1,APELIDO_2,LOCALIDADE,SALARIO,DATA_NACEMENTO,\n" +
+                    "            (\n" +
+                    "            select NOME\n" +
+                    "                from empregado\n" +
+                    "                where nss = e1.NSS_SUPERVISA\n" +
+                    "           ) as nome_xefe,\n" +
+                    "            (\n" +
+                    "                select NOME_DEPARTAMENTO\n" +
+                    "                from departamento\n" +
+                    "                where NUM_DEPARTAMENTO = e1.NUM_DEPARTAMENTO_PERTENECE\n" +
+                    "                ) as departamento\n" +
+                    "    from empregado as e1\n" +
+                    "    where LOCALIDADE = '"+localidad+"'");
+            if(!rs.next()){
+                System.out.println("No hay empleados registrados de: "+localidad);
+            }else{
+                System.out.println("Los empleados de "+localidad+" registrados son: ");
+                while (rs.next()){
+                    String nome = rs.getString("NOME");
+                    String apelido1 = rs.getString("APELIDO_1");
+                    String apelido2 = rs.getString("APELIDO_2");
+                    String localidade = rs.getString("LOCALIDADE");
+                    double salario = rs.getDouble("SALARIO");
+                    Date dataNacemento = rs.getDate("DATA_NACEMENTO");
+                    String nomeXefe = rs.getString("nome_xefe");
+                    String departamento = rs.getString("departamento");
+
+                    System.out.println("-".repeat(15));
+                    System.out.println("Nombre: " + nome);
+                    System.out.println("Apellido 1: " + apelido1);
+                    System.out.println("Apellido 2: " + apelido2);
+                    System.out.println("Localidad: " + localidade);
+                    System.out.println("Salario: " + salario);
+                    System.out.println("Fecha de Nacimiento: " + dataNacemento);
+                    System.out.println("Nombre del Jefe: " + (nomeXefe != null ? nomeXefe : "Sin jefe"));
+                    System.out.println("Departamento: " + (departamento != null ? departamento : "Sin departamento"));
+                    System.out.println("-".repeat(15));
+                }
+            }
+            rs.close();
+        }catch (SQLException e){
+            System.out.println("Error al consultar empleados por localidad.");
             e.printStackTrace();
         }
     }
+
+
 
 
     public static void desconectarDB(){
